Fix UpdateNumber causing crashes

+++ src/protocol/Edelstein.Protocol.Gameplay/Entities/Inventories/Modifiers/Operations/StructuredModifyInventoryOperationInfoUpdateNumber.cs
@@ -5,5 +5,5 @@ namespace Edelstein.Protocol.Gameplay.Entities.Inventories.Modifiers.Operations;
 public record StructuredModifyInventoryOperationInfoUpdateNumber : StructuredModifyInventoryOperationInfo
 {
     [FieldOrder(0)]
-    public required int Number { get; init; }
+    public required short Number { get; init; }
 }
